requirements.txt
wget
python-dotenv